:sparkles: create useMineSweeper
